feat(live pnl): create a live price hook for live PnL
