+++ src/components/organisms/ChainOtcStateDetails/ChainOtcStateDetails.module.scss
@@ -1,0 +1,87 @@
+@import 'styles/variables.scss';
+
+.cards {
+	display: flex;
+
+	.box {
+		max-width: 620px;
+		width: 310px;
+		background-color: var(--color-white);
+		border-radius: 10px;
+		padding: 20px;
+
+		.toggle {
+			float: right;
+			margin-bottom: 8px;
+			color: var(--color-text-disabled);
+			font-size: 14px;
+
+			svg {
+				margin-bottom: -4px;
+				margin-left: var(--space-4);
+			}
+		}
+		.toggle:hover {
+			cursor: pointer;
+			color: var(--color-primary);
+		}
+
+		.content {
+			p {
+				font-weight: 600;
+			}
+			.column {
+				display: flex;
+				justify-content: space-between;
+			}
+		}
+	}
+
+	/* Simulator */
+	.simulator {
+		border-top-right-radius: 10px;
+		border-bottom-right-radius: 10px;
+	}
+
+	.changeEdge {
+		border-top-right-radius: 0;
+	}
+
+	.hide {
+		display: none;
+	}
+
+	.show {
+		display: block;
+	}
+
+	/* ------------------ */
+}
+
+.funded {
+	display: flex;
+	justify-content: center;
+
+	:first-child {
+		margin-right: 10px;
+	}
+
+	svg {
+		margin-top: -4px;
+	}
+}
+
+.buttons {
+	display: flex;
+	margin-top: 20px;
+}
+
+.disabled {
+	color: var(--color-text-disabled);
+}
+
+.expirations {
+	display: flex;
+	justify-content: space-between;
+	color: var(--color-text-disabled);
+}
